refactor(): correct memory typo
